bug: code-connect was creating icon names in kebab-case. Changed to snake_case bug: customTypes wasn't being exported. Moved to root. chore: updated all devDependencies to latest version
