+++ package.json
@@ -8,6 +8,7 @@
   "files": [
     "index.css",
     "dist/index.*s",
+    "dist/types.*s",
     "dist/outputs",
     "outputs/web",
     "outputs/icons",
@@ -22,28 +23,28 @@
     "./index.css": "./index.css"
   },
   "devDependencies": {
-    "@actions/core": "^1.10.0",
-    "@figma/code-connect": "^1.1.4",
-    "@types/chai": "^4.3.16",
+    "@actions/core": "^1.11.1",
+    "@figma/code-connect": "^1.2.2",
+    "@types/chai": "^5.0.1",
     "@types/md5": "^2.3.5",
-    "@types/mocha": "^10.0.6",
-    "@types/node": "^20.12.7",
-    "@types/react": "^18.3.11",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^22.10.1",
+    "@types/react": "^18.3.12",
     "@types/sinon": "^17.0.3",
-    "chai": "^5.1.0",
+    "chai": "^5.1.2",
     "cheerio": "^1.0.0",
-    "dotenv": "^16.4.5",
+    "dotenv": "^16.4.7",
     "jest": "^29.7.0",
     "md5": "^2.3.0",
-    "mocha": "^10.4.0",
-    "oslllo-svg-fixer": "^3.0.0",
-    "prettier": "^3.3.3",
+    "mocha": "^11.0.1",
+    "oslllo-svg-fixer": "^5.0.0",
+    "prettier": "^3.4.2",
     "scale-that-svg": "^1.0.6",
-    "sharp": "^0.33.3",
-    "sinon": "^17.0.1",
-    "ts-jest": "^29.1.2",
+    "sharp": "^0.33.5",
+    "sinon": "^19.0.2",
+    "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
-    "typescript": "^5.4.5",
+    "typescript": "^5.7.2",
     "webfont": "^11.2.26",
     "zeta-icon-name-checker": "^0.0.15"
   },
