+++ scripts/code-connect/generate-code-connect-call.ts
@@ -76,7 +76,7 @@ const getTemplate = (iconName: string) =>
               ts.factory.createIdentifier("props"),
               ts.factory.createIdentifier("rounded"),
             ),
-            ts.factory.createTemplateTail(`>${iconName.toLowerCase().replaceAll(" ", "-")}</zeta-icon>`),
+            ts.factory.createTemplateTail(`>${iconName.toLowerCase().replaceAll(" ", "_")}</zeta-icon>`),
           ),
         ]),
       ),
