+++ scripts/fetch-icons/generators/generateFonts.ts
@@ -1,5 +1,5 @@
 import { Result } from "webfont/dist/src/types/Result.js";
-import { FontType, GenerateFontResult } from "../../types/customTypes.js";
+import { FontType, GenerateFontResult } from "../../../types.js";
 import { webfont } from "webfont";
 import { writeFileSync } from "fs";
 import { createFolder, getIconFileName, toSnakeCase } from "../../utils/fileUtils.js";
@@ -121,9 +121,8 @@ function getDartIconDefinition(iconName: string, unicode: string, type: FontType
   }
 
   return `${iconPreview}
-  static const IconData ${iconName} = IconData(0x${unicode}, fontFamily: family${
-    type?.capitalize() ?? ""
-  }, fontPackage: package);`;
+  static const IconData ${iconName} = IconData(0x${unicode}, fontFamily: family${type?.capitalize() ?? ""
+    }, fontPackage: package);`;
 }
 
 function getIconPreview(iconName: string, type: FontType | undefined) {
@@ -132,7 +131,6 @@ function getIconPreview(iconName: string, type: FontType | undefined) {
     .split("_")
     .map((e) => e.capitalize())
     .join(" ");
-  return `  /// <i> <img width='48' src="${GITHUB_URL}${name_link}.png"></br>${readableName} icon ${
-    type != undefined ? `(${type})` : ""
-  }</i>`;
+  return `  /// <i> <img width='48' src="${GITHUB_URL}${name_link}.png"></br>${readableName} icon ${type != undefined ? `(${type})` : ""
+    }</i>`;
 }
